Model builder now has *podunk* pagination support
* The pagesize and currentpage properties on the input pagination object control the results returned.
* Additionally the total count of possible records is returned.

SDK Generics currently limit the ability to paginate results.
We are using maximum rows to limit the results and then taking the page
out of the end of the results. If proper pagination were to be added to Dovetail SDK
this interface should not have to change.

*** Suggestion ***

Don't get the last page of 1 million rows. Get the first page of the same query sorted in the opposite direction.