History updater overload gets an instance of act entry template
We copy the act entry template before processing the history item. Any changes to the act entry template are within the context of the item being processed. The idea behind this is to allow the delegate to modify the template being executed. For example if the template decides to return HTML it can squelch the Htmlizer action with it's own new action.